Add simple loading in events,event and profile, load data client side, extend api endpoints

+++ pages/api/events/[id]/index.ts
@@ -1,27 +1,52 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import prisma from '../../../../lib/prisma';
+import { getServerSession } from 'next-auth/next';
+import { authOptions } from '../../auth/[...nextauth]';
 
 export default async function handler(
     req: NextApiRequest,
     res: NextApiResponse
 ) {
-    // TODO: check authentication
-    // DELETE api/events/{id}
-    if (req.method === 'DELETE') {
+    const session = await getServerSession(req, res, authOptions);
+    if (session) {
         try {
-            const eventId = req.query.id.toString();
-            const event = await prisma.event.delete({
-                where: { id: eventId },
-            });
-            res.status(200).json({ message: 'Deleted event successfully' });
+            // DELETE api/events/{id}
+            if (req.method === 'DELETE') {
+                const eventId = req.query.id.toString();
+                const event = await prisma.event.delete({
+                    where: { id: eventId },
+                });
+                res.status(200).json({ message: 'Deleted event successfully' });
+            }
+            // GET event detail api/events/{id}
+            else if (req.method === 'GET') {
+                const event = await prisma.event.findUnique({
+                    where: {
+                        id: String(req.query.id),
+                    },
+                    include: {
+                        host: {
+                            select: { firstName: true },
+                        },
+                        menu: {
+                            select: {
+                                id: true,
+                                title: true,
+                                description: true,
+                                link: true,
+                            },
+                        },
+                    },
+                });
+                res.status(200).json({ event: event });
+            } else {
+                res.status(405).end('Method Not Allowed');
+            }
         } catch (err) {
-            res.status(500).json({ statusCode: 500, message: err.message });
+            res.status(500).json({ message: err.message });
         }
     } else {
-        res.status(200).json({
-            statusCode: 200,
-            info: 'TODO',
-        });
+        res.status(401).json({ message: 'Not authenticated' });
     }
 }
